Consolidate SSA tests into a single file
Merge serverside_apply_kinds_test.go back into serverside_apply_test.go.
The split was an artifact of how the tests were written, not a deliberate
organizational choice.
